+++ wsidicomizer/sources/opentile/opentile_metadata.py
@@ -61,7 +61,9 @@ class OpenTileMetadata(WsiDicomizerMetadata):
         )
         if objectives is not None or icc_profile is not None:
             optical_paths = [
-                OpticalPath("1", objective=objectives, icc_profile=icc_profile)
+                OpticalPath(
+                    "1", objective=objectives, icc_profile=icc_profile
+                ).add_color_space_from_icc()
             ]
         else:
             optical_paths = []